third commit

+++ project3/prelim1test.c
@@ -2,13 +2,13 @@
 
 int main(){
 
-    int num1,num2;
+    float num1,num2;
 
 
-    scanf("%d", &num1);
-    scanf("%d", &num2);
-    int result = num1 + num2;
-    printf("%d",result);
+    scanf("%f", &num1);
+    scanf("%f", &num2);
+    float result = num1 + num2;
+    printf("%f",result);
 
     return 0;
